// src/redux/summary/types.ts
export enum ActionType {
  LOAD_HOME = 'SUMMARY/LOAD_HOME_STATISTICS',
  LOAD_HOME_SUCCESS = 'SUMMARY/LOAD_HOME_SUCCESS',
  LOAD_COMMISSION_TODAY = 'SUMMARY/LOAD_COMMISSION_TODAY',
  LOAD_COMMISSION_TODAY_SUCCESS = 'SUMMARY/LOAD_COMMISSION_TODAY_SUCCESS',
  LOAD_COMMISSION_EXPECT = 'SUMMARY/LOAD_COMMISSION_EXPECT',
  LOAD_COMMISSION_EXPECT_SUCCESS = 'SUMMARY/LOAD_COMMISSION_EXPECT_SUCCESS',
  LOAD_SPU_SALES_TOP = 'SUMMARY/LOAD_SPU_SALES_TOP',
  LOAD_SPU_SALES_TOP_SUCCESS = 'SUMMARY/LOAD_SPU_SALES_TOP_SUCCESS',
  LOAD_SPU_COMMISSION_TOP = 'SUMMARY/LOAD_SPU_COMMISSION_TOP',
  LOAD_SPU_COMMISSION_TOP_SUCCESS = 'SUMMARY/LOAD_SPU_COMMISSION_TOP_SUCCESS',
  LOAD_COMMISSION_HISTORY = 'SUMMARY/LOAD_COMMISSION_HISTORY',
  LOAD_COMMISSION_HISTORY_SUCCESS = 'SUMMARY/LOAD_COMMISSION_HISTORY_SUCCESS',
}